Fixed tests, including moving helper methods to reduce config dependencies when running unit tests
